requirements.txt
appdirs==1.4.4
astor==0.8.1
attrs==23.1.0
Automat==22.10.0
CacheControl==0.13.1
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.1.1
cycler==0.12.1
docutils==0.20.1
filelock==3.13.1
fonttools==4.43.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
iniconfig==2.0.0
kiwisolver==1.4.5
lunr==0.6.2
matplotlib==3.8.0
mock==5.1.0
msgpack==1.0.7
numpy==1.26.0
packaging==23.2
patch==1.16
Pillow==10.0.1
pluggy==1.3.0
pydoctor==23.9.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
toml==0.10.2
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
urllib3==1.26.18
zope.interface==6.1
